Minor changes i did to core files. As for cart.cpp, i had to fix the CRC values so we now use hex values as it should be.

+++ src/handy-libretro/machine.h
@@ -57,8 +57,8 @@ typedef int16_t SWORD;
 typedef uint16_t UWORD;
 
 // Longs should be 32-bits wide
-typedef signed long SLONG;
-typedef unsigned long ULONG;
+typedef int32_t SLONG;
+typedef uint32_t ULONG;
 
 // Read/Write Cycle definitions
 #define CPU_RDWR_CYC	5
